off_chain/requirement.txt
Flask
psycopg2
python-dotenv
PyQt5
PyQt5-sip